Refactor obey Entity classed inheritance for toJson
