+++ modules/standard/src/rescuecore2/standard/entities/AmbulanceCentre.java
@@ -62,9 +62,7 @@ public class AmbulanceCentre extends Building {
 
   @Override
   public JSONObject toJson() {
-    JSONObject jsonObject = super.toJson();
-    jsonObject.put( "EntityName", this.getEntityName() );
-
-    return jsonObject;
+    return super.toJson();
   }
+
 }
