+++ modules/standard/src/rescuecore2/standard/entities/AmbulanceTeam.java
@@ -51,9 +51,7 @@ public class AmbulanceTeam extends Human {
 
   @Override
   public JSONObject toJson() {
-    JSONObject jsonObject = super.toJson();
-    jsonObject.put( "EntityName", this.getEntityName() );
-
-    return jsonObject;
+    return super.toJson();
   }
+
 }
