+++ modules/standard/src/rescuecore2/standard/entities/Civilian.java
@@ -51,9 +51,8 @@ public class Civilian extends Human {
 
   @Override
   public JSONObject toJson() {
-    JSONObject jsonObject = super.toJson();
-    jsonObject.put( "EntityName", this.getEntityName() );
-
-    return jsonObject;
+    return super.toJson();
   }
+
+
 }
