+++ modules/standard/src/rescuecore2/standard/entities/FireBrigade.java
@@ -123,12 +123,9 @@ public class FireBrigade extends Human {
     return "Fire brigade";
   }
 
-
   @Override
   public JSONObject toJson() {
-    JSONObject jsonObject = super.toJson();
-    jsonObject.put( "EntityName", this.getEntityName() );
-
-    return jsonObject;
+    return super.toJson();
   }
+
 }
