+++ modules/standard/src/rescuecore2/standard/entities/FireStation.java
@@ -62,8 +62,8 @@ public class FireStation extends Building {
 
   @Override
   public JSONObject toJson() {
-    JSONObject jsonObject = super.toJson();
-    jsonObject.put( "EntityName", this.getEntityName() );
-    return jsonObject;
+    return super.toJson();
   }
+
+
 }
