+++ modules/standard/src/rescuecore2/standard/entities/GasStation.java
@@ -36,8 +36,7 @@ public class GasStation extends Building {
 
   @Override
   public JSONObject toJson() {
-    JSONObject jsonObject = super.toJson();
-    jsonObject.put( "EntityName", this.getEntityName() );
-    return jsonObject;
+    return super.toJson();
   }
+
 }
