+++ modules/standard/src/rescuecore2/standard/entities/Hydrant.java
@@ -47,8 +47,7 @@ public class Hydrant extends Road {
 
   @Override
   public JSONObject toJson() {
-    JSONObject jsonObject = super.toJson();
-    jsonObject.put( "EntityName", this.getEntityName() );
-    return jsonObject;
+    return super.toJson();
   }
+
 }
