+++ modules/standard/src/rescuecore2/standard/entities/PoliceOffice.java
@@ -62,8 +62,7 @@ public class PoliceOffice extends Building {
 
   @Override
   public JSONObject toJson() {
-    JSONObject jsonObject = super.toJson();
-    jsonObject.put( "EntityName", this.getEntityName() );
-    return jsonObject;
+    return super.toJson();
   }
+
 }
