+++ modules/standard/src/rescuecore2/standard/entities/Refuge.java
@@ -63,8 +63,7 @@ public class Refuge extends Building {
 
   @Override
   public JSONObject toJson() {
-    JSONObject jsonObject = super.toJson();
-    jsonObject.put( "EntityName", this.getEntityName() );
-    return jsonObject;
+    return super.toJson();
   }
+
 }
